Create My Travel Wishlist
In this HTML document, I created a simple webpage titled "My Travel Wishlist." The page showcases a list of places I would love to visit, each accompanied by an image and a brief description. The CSS styling was applied to format the content, including centering the heading, aligning the images and text, and adding box shadows on hover to create a smooth visual effect. I expressed my desire to explore the unique cultural experiences, landmarks, and natural beauty offered by each destination, from England's celebrated arts and sciences to Paris, the iconic City of Love. The webpage serves as a glimpse into my travel aspirations and highlights the beauty of diverse destinations around the world.